Make staging image coordinates as unique as possible
It is a good idea to keep the staging image locations as unique
as possible. This avoids trouble when the imagePullPolicy of the
gitlab runner for the jobs is not set to 'always' (which is a
dangerous setting but may occur). Add two new behaviors in this
commit:
1) If somebody explicitly defines repository AND tag for the
staging location, print a warning in the log file.
2) If somebody defines the repository explicitly, and replaces
the auto calculated git revision with a static value, then
append the first 7 chars of the git commit to the tag (which is
by default the image id).