Add update check logic for integration test
add first gipgee-update-check command for alpine containers.

Later, we can embed this in gipgee as preconfigured upgrade check
for alpine based containers and extend it with support for debian
or redhat based distros.

Add a few code changes to make the update check working (gipgee
executes the update check command)