Kanikobuild: increase alpine compatibility
alpine container build fails after apk upgrade,
seems to be a problem with /var/run which is described
in https://github.com/GoogleContainerTools/kaniko/issues/1297.

A workaround is to move /var away and add the new location to
the ignored paths. For that purpose, implement a quirks: section
in the config which allows to toggle this quirk for kaniko builds.