Pass the image id as build arg in the image build pipeline
pass the image id of the image to build as build arg via kaniko. Build
args name is GIPGEE_IMAGE_ID. This can be useful if you want to use
the same container file for different images and need to detect
the current image id in your container image preparation scripts.