Append the image id to the staging test command
the image staging command needs to know the image id, too. Just append
it as last parameter.

This is useful if you want to implement slightly different tests for the
same image family in the same test file.

If someone wants to call a test command which doesn't like this
additional param, he can simply wrap it with a small shell script that
consumes the last parameter and removes it before calling the desired
test script.