Add quirk for manual generated pipelines
when a child pipeline only contains jobs that have
to be triggered manually and the parent pipeline runs
with "strategy: depend", gitlab raises errors for the
parent pipeline or keeps the parent pipeline in the running
state. So we just add a "make gitlab happy" job which always
runs in the generated pipeline.

This is a preparation for an "universal" generate pipeline command
which detects the start conditions of the pipeline (image rebuild in
feature branch, image (re)build in default branch, update check) and
then generates the corresponding pipeline. The .gitlab-ci.yml
then only has to run this universal gipgee command and trigger the
generated pipeline, which keeps the content of the .gitlab-ci.yml small.