Add non root container images
Add non root container images with update check command []. As
preperation to be able to test the parent child image detection of
gipgee (which has to be implemented).

Improve the update check pipeline so that it checks for the
update check command length instead of nil. nil means, that there
is no command defined (even if it's explicitly there with null value
in yaml), so define [] as default if users want to disable the update
check command.