First step of working skopeo inspect location
fix (default) base image credentials, remove the
defaultBaseImageRegistryCredentials field in the defaults section
because there is a defaultBaseImage section which can and should be used
for that. Implement skopeo layer inspect which downloads the layers
infos of the base image and the release locations, so that I can
implement the layer comparison (checking if base image layers are still
the first layers of the released images).