Implement Skopeo update check and result check
implement skopeo update check completely, compare base image layers
and current image layers and, if diverged, register for rebuild.

Implement job that collects the skopeo and update check results at the
end of the update checks. This job currently prints out the images to be
rebuild, but does not trigger a rebuild pipeline - yet. Thats the next
step of the implementation that will follow.