Fix bug in container image coordinate marshalling
I thought it would be a good idea to handle container image
coordinates as urls - but they are not valid urls because they
don't contain a scheme at the beginning of the coordinates.

This caused an error in the ContainerImageCoordinates yaml marshalling.
Implemented a unit test to reproduce the bug and then fix it by just
using filepath.join instead of using the url methods in the marshalling
method.