Set DOCKER_AUTH_CONFIG per job in self release pipeline
there seems to be a bug that if the first child pipeline contains
another "global" variable DOCKER_AUTH_CONFIG, that it overrides
that globally set DOCKER_AUTH_CONFIG of the second child pipeline.

I don't know if this is intentionally done by gitlab, but in this
case just scope the DOCKER_AUTH_CONFIG to the corresponding pipeline
jobs in the self release pipeline instead of setting it globally - so
that it does not disturb the image build pipeline which will be
generated for testing purpose in the self release pipeline.

In the released gipgee, the docker auth will always be the same for
update check / image build pipelines, so that this shouldn't be a
problem then.