Add go-zglob dependency
for the feature branch delta detection, we need a glob like
syntax in the config file which can be matched against the
current diff between a feature and the default branch in gitlab.

Because golangs internal file path match is broken
(https://github.com/golang/go/issues/11862) is seems to be necessary
to use an external lib. go-zglob does this very well, so use it.